requirement.txt
python-dotenv
python-telegram-bot
openai 